refactor(openclaw-plugin): extract lib/openclaw, lib/delivery, lib/utils
Moves plugin-api.ts → lib/openclaw/, delivery.dispatcher.ts + delivery.prompt.ts → lib/delivery/, sanitize.ts → lib/utils/. Updates all import paths in index.ts, prompt files, and test files. No logic changes.
