+++ packages/openclaw-plugin/src/lib/openclaw/plugin-api.ts
