+++ packages/openclaw-plugin/src/lib/utils/sanitize.ts
